refactored plugin for order id mask creation
