+++ app/code/Deity/Sales/etc/di.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0"?>
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
+    <type name="Magento\Sales\Api\OrderManagementInterface">
+        <plugin name="deity_magento_order_management" type="Deity\Sales\Plugin\Sales\Api\OrderManagement" />
+    </type>
+</config>
